Committing what I have so far
Testing indicates that this will take some more time than I have at this
time.  I was hoping to get things done while on this trip, but this trip
is taking a whole lot out of me.

The minimum spanning tree sort of works, only thing is that it does not
do a backtrack to see if any city was missed.
So, for example, any city that has hyderabad as the shortest path, would
end up in cape town, and the algorrithm would end there.

I really need your help on this one, Daniel