Comment added
The comments are updated and file formatted.
It looks good, and it would seem like it does not matter which city we
start, we have the same list, regardless.
I will try to look at the output to compare and make sure.

I think this is due to the fact we are using somewhat of a what was the
second algorithm?